requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
chromadb>=0.4.0
python-jose>=3.3.0
cryptography>=3.4.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
httpx>=0.24.0
pytest-asyncio>=0.21.0
